Convert universal time <-> greenwich sidereal time.

+++ rust-implementation/src/lib/macros.rs
@@ -3,7 +3,7 @@ use crate::lib::util as utils;
 /// Convert a Civil Time (hours,minutes,seconds) to Decimal Hours
 ///
 /// Original macro name: HMSDH
-pub fn hms_dh(hours: u32, minutes: u32, seconds: u32) -> f64 {
+pub fn hms_dh(hours: u32, minutes: u32, seconds: f64) -> f64 {
     let f_hours = hours as f64;
     let f_minutes = minutes as f64;
     let f_seconds = seconds as f64;
@@ -51,13 +51,13 @@ pub fn dh_min(decimal_hours: f64) -> u32 {
 /// Return the seconds part of a Decimal Hours
 ///
 /// Original macro name: DHSec
-pub fn dh_sec(decimal_hours: f64) -> u32 {
+pub fn dh_sec(decimal_hours: f64) -> f64 {
     let a = decimal_hours.abs();
     let b = a * 3600.0;
     let c = utils::round_f64(b - 60.0 * (b / 60.0).floor(), 2);
     let d = if c == 60.0 { 0.0 } else { c };
 
-    return d as u32;
+    return d;
 }
 
 /// Convert a Greenwich Date/Civil Date (day,month,year) to Julian Date
